+++ modules/admin/src/ngrest/plugins/CheckboxRelation.php
@@ -16,7 +16,10 @@ use admin\ngrest\base\Model;
  * 
  * public $groups = [];
  * 
- * public $extraFields = ['groups'];
+ * public function extraFields()
+ * {
+ *     return ['groups'];
+ * }
  * 
  * ```
  *